+++ src/Libraries/SmartStore.Services/DataExchange/Export/DataExporter.cs
@@ -686,7 +686,7 @@ namespace SmartStore.Services.DataExchange.Export
 							OrderBy = ProductSortingEnum.Position,
 							PageSize = int.MaxValue,
 							StoreId = (ctx.Request.Profile.PerStore ? ctx.Store.Id : ctx.Filter.StoreId),
-							VisibleIndividuallyOnly = true,
+							VisibleIndividuallyOnly = ctx.Projection.OnlyIndividuallyVisibleAssociated,
 							ParentGroupedProductId = product.Id
 						};