+++ src/Libraries/SmartStore.Services/DataExchange/Export/ExportProfileService.cs
@@ -116,6 +116,7 @@ namespace SmartStore.Services.DataExchange.Export
 						CriticalCharacters = "¼,½,¾",
 						PriceType = PriceDisplayType.PreSelectedPrice,
 						NoGroupedProducts = (features.HasFlag(ExportFeatures.CanOmitGroupedProducts) ? true : false),
+						OnlyIndividuallyVisibleAssociated = true,
 						DescriptionMerging = ExportDescriptionMerging.Description
 					};